Added 7-print_last_digit.c source file

+++ 0x02-functions_nested_loops/main.h
@@ -15,6 +15,6 @@ int print_sign(int n);
 
 int _abs(int n);
 
-int print_last_digit(int n);
+int print_last_digit(long int n);
 
 #endif
